builder/requirements.txt
langchain
neo4j
neomodel
requests
python-dotenv
uvicorn
fastapi
langchain_ollama
confluent-kafka